requirements.txt
# Python依赖包
# 安装方法: pip install -r requirements.txt

# 加密库
pycryptodome>=3.19.0    # RC4加密
gmssl>=3.2.3            # SM3国密算法

# HTTP请求
requests>=2.31.0        # HTTP客户端

# 可选: 如果需要更好的性能
# pycryptodomex>=3.19.0
